+++ src/views/boardlist.html
@@ -24,6 +24,7 @@
                 <tr>
                     <th>글번호</th>
                     <th>제목</th>
+                    <th>작성자</th>
                 </tr>
             </thead>
             <tbody id="boardList"></tbody>
@@ -36,13 +37,13 @@
             <button type="submit">검색</button>
         </form>
         
-        <button type="button" onclick="location.href='/write'">글쓰기</button>
+        <button type="button" id="writebtn">글쓰기</button>
     </div>
     <script>
         fetch('/header.html')
             .then(response => response.text())
-            .then(data => {
-                document.getElementById('header').innerHTML = data;
+            .then(header => {
+                document.getElementById('header').innerHTML = header;
                 // header.html 내의 스크립트 실행
                 const scripts = document.getElementById('header').getElementsByTagName('script');
                 for (let script of scripts) {
@@ -50,10 +51,8 @@
                 }
             })
             .catch(error => console.error('Error loading header:', error));
-
-        let currentPage = 1;
-        const pagelimit = 5;
-        const postlimit = 20;
+            
+        const currentPage = 1;
         let searchKeyword="";
 
         //검색 기능 추가
@@ -62,9 +61,16 @@
             searchKeyword = document.getElementById('searchKeyword').value;
             renderPagination(currentPage);
         });
+
+        //글쓰기버튼 기능
+        document.getElementById('writebtn').addEventListener('click',function(event){
+                window.location.href = '/write';
+        });
         
         // 게시물과 페이지 버튼을 렌더링하는 메인 함수
         async function renderPagination(page) {
+            const pagelimit = 5;
+            const postlimit = 20;
             const response = await fetch(`/api/board/list?page=${page}&pageSize=${postlimit}&searchKeyword=${encodeURIComponent(searchKeyword)}`);
             const {boardData,totalPage} = await response.json();
             
@@ -75,6 +81,7 @@
                 <tr>
                     <td>${board.id}</td>
                     <td><a href="/view/${board.id}">${board.title}</a></td>
+                    <td>${board.email}</td>
                 </tr>
             `)
             .join('');
@@ -125,7 +132,7 @@
             
         }
 
-        renderPagination(currentPage); // 첫 페이지 로드 시 호출
+        renderPagination(currentPage); // 첫 페이지 로드 시 호출    
     </script>
 </body>
 </html>
