+++ src/views/boardmodify.html
@@ -39,7 +39,6 @@
         fetch('/header.html')
             .then(response => response.text())
             .then(data => {
-                console.log(data);
                 document.getElementById('header').innerHTML = data;
                 // header.html 내의 스크립트 실행
                 const scripts = document.getElementById('header').getElementsByTagName('script');
@@ -48,22 +47,39 @@
                 }
             })
             .catch(error => console.error('Error loading header:', error));
+
+        //로그인 체크
+        const token = sessionStorage.getItem("authToken");
+        if (!token){
+            window.location.href = '/login';
+            alert("로그인 후에 사용할 수 있는 기능입니다.");
+        }
         
         // 현재 게시물 ID를 URL에서 추출
-        const currentUrl = window.location.pathname;
-        const postId = currentUrl.split('/').pop(); // /board/update/26 에서 26 추출
+        const postId = window.location.pathname.split('/').pop(); // URL에서 postId 추출
 
-        // API 호출하여 게시물 데이터 가져오기
-        fetch(`/api/board/post/${postId}`)
-            .then(res => res.json())
-            .then(data => {
-                // 제목과 내용을 폼에 미리 채워넣기
-                document.getElementById('title').value = data.title;
-                document.getElementById('content').value = data.content;
-            })
-            .catch(error => {
-                console.error('게시물 데이터를 불러오는데 실패했습니다:', error);
-            });
+        async function fetchPostData(){
+            try {
+                // 게시물 데이터 가져오기
+                const postData = await (await fetch(`/api/board/post/${postId}`)).json();
+                
+                // 사용자 이메일 확인 및 권한 검사
+                const tokenEmail = await (await fetch(`/api/user/tokenToEmail`, {
+                    method: 'POST',
+                    headers: { 'authorization': token },
+                })).json();
+
+                if (tokenEmail !== postData.email) {
+                    window.history.back();
+                    alert('게시글 수정에 대한 권한이 없습니다.');
+                }
+                
+                document.getElementById('title').value = postData.title;
+                document.getElementById('content').value = postData.content;
+            } catch (error) {
+                console.error('오류 발생:', error);
+            }
+        };
 
         // 폼 제출 이벤트 처리
         document.getElementById('updateForm').addEventListener('submit', function(event) {
@@ -93,6 +109,8 @@
             })
             .catch(error => console.error('Error:', error))
         });
+
+        fetchPostData();
     </script>
 </body>
 </html>
