+++ src/views/boardwrite.html
@@ -50,6 +50,13 @@
             })
             .catch(error => console.error('Error loading header:', error));
         
+        //로그인 체크
+        const token = sessionStorage.getItem("authToken");
+        if (!token){
+            window.location.href = '/login';
+            alert("로그인 후에 사용할 수 있는 기능입니다.");
+        }
+    
         document.getElementById('postForm').addEventListener('submit',function(event){
             event.preventDefault();
 
@@ -60,6 +67,7 @@
                 method: 'POST',
                 headers: {
                     'Content-Type': 'application/json', // JSON 형식으로 데이터 전송
+                    'authorization': token,
                 },
                 body: JSON.stringify({title,content}),
             })
